refactor: Updated response and error handlers
