+++ src/api/middlewares/errorHandler.ts
@@ -1,12 +1,7 @@
-import { CustomRequest, CustomResponse, CustomNext } from "@/types/request";
-import { appConfig } from "@/config";
+import { CustomRequest, CustomResponse, CustomNext } from "@/types/route";
 import { createNewLog } from "@/utils/logger";
 
 export const errorHandler = (err: any, req: CustomRequest, res: CustomResponse, next: CustomNext) => {
     createNewLog(err, req, res, next);
-    res.error(
-        err,
-        appConfig.env === "development" ? err.message : "Internal Server Error.",
-        err?.status || 500
-    );
+    res.error(err);
 };
